Move to OverviewPage, some code improvements from review
